# pkg/models/types.gen.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version (devel) DO NOT EDIT.
package models

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

// CreatedSubmission defines model for CreatedSubmission.
type CreatedSubmission struct {
	Id   SubmissionId `json:"id"`
	Link *string      `json:"link,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
}

// JudjeStatus defines model for JudjeStatus.
type JudjeStatus = string

// Problem defines model for Problem.
type Problem struct {
	Id        *ProblemId `json:"id,omitempty"`
	Name      *string    `json:"name,omitempty"`
	Statement string     `json:"statement"`
}

// ProblemId defines model for ProblemId.
type ProblemId = openapi_types.UUID

// Problems defines model for Problems.
type Problems struct {
	Items []struct {
		Id     ProblemId   `json:"id"`
		Link   string      `json:"link"`
		Name   string      `json:"name"`
		Status JudjeStatus `json:"status"`
	} `json:"items"`
}

// Submission defines model for Submission.
type Submission struct {
	Id         SubmissionId       `json:"id"`
	Language   SubmissionLanguage `json:"language"`
	ProblemId  ProblemId          `json:"problemId"`
	SourceCode string             `json:"sourceCode"`
	Status     JudjeStatus        `json:"status"`
}

// SubmissionId defines model for SubmissionId.
type SubmissionId = openapi_types.UUID

// SubmissionLanguage defines model for SubmissionLanguage.
type SubmissionLanguage = string

// SubmissionStatus defines model for SubmissionStatus.
type SubmissionStatus struct {
	Id        SubmissionId `json:"id"`
	ProblemId ProblemId    `json:"problemId"`
	Status    JudjeStatus  `json:"status"`
}

// SubmitBody defines model for SubmitBody.
type SubmitBody struct {
	Language   SubmissionLanguage `json:"language"`
	ProblemId  ProblemId          `json:"problemId"`
	SourceCode string             `json:"sourceCode"`
}

// ProblemIdParameter defines model for ProblemIdParameter.
type ProblemIdParameter = ProblemId

// SubmissionIdParameter defines model for SubmissionIdParameter.
type SubmissionIdParameter = SubmissionId

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse = Error

// SumbitSolutionJSONRequestBody defines body for SumbitSolution for application/json ContentType.
type SumbitSolutionJSONRequestBody = SubmitBody